Efficient Influence Function Computer

compute_eif() computes the efficient influence function for the chosen parameter using the already estimated nuisance parameters. If certain nuisance parameters are known, such as propensity scores in a randomized control trial, then they may be input directly into this function. Note that this EIF is only uncentered in expectation; it should only be used with the one-step estimator. When ace_estimate is NULL, and pseudo-uncentered EIF is computed. It should only be used with the one-step estimator, as it is only uncentered in expectation.

Usage

compute_eif(
  data,
  effect,
  confounders,
  exposure,
  outcome,
  modifiers,
  prop_score_fit,
  prop_score_values,
  cond_outcome_fit,
  failure_hazard_fit,
  censoring_hazard_fit,
  ace_estimate,
  plugin_estimates
)

Arguments

data

A data.table containing the observed data. train_data is formatted by unihtee().

effect

A character indicating the type of treatment effect modifier variable importance parameter. Currently supports "absolute" and "relative".

confounders

A character vector of column names corresponding to baseline covariates.

exposure

A character corresponding to the exposure variable.

outcome

A character corresponding to the outcome variable.

modifiers

A character vector of columns names corresponding to the suspected effect modifiers. This vector must be a subset of confounders.

prop_score_fit

A list output by the fit_prop_score() function.

prop_score_values

A numeric vector corresponding to the (known) propensity score values for each observation in data.

cond_outcome_fit

A list output by the fit_failure_hazard() function.

failure_hazard_fit

A list output by the fit_cond_outcome() function.

censoring_hazard_fit

A list output by the fit_censoring_hazard() function.

ace_estimate

A numeric estimate of the average causal effect associated with the specified effect type and data-generating process.

plugin_estimates

A numeric vector of plug-in estimate of the treatment effect modifier variable importance parameters associated with the specified effect type and data-generating process.

Value

A data.table whose columns are the efficient influence functions of each variable in modifiers. The rows correspond to the observations of data.