Cross-Fitting Procedure

cross_fit_fold() estimates the treatment effect modification variable importance parameters using cross-fitting.

cross_fit_fold(
  fold,
  data,
  confounders,
  modifiers,
  exposure,
  outcome,
  censoring,
  outcome_type,
  effect,
  estimator,
  cond_outcome_estimator,
  prop_score_estimator,
  prop_score_values,
  failure_hazard_estimator,
  censoring_hazard_estimator
)

Arguments

fold

An origami fold object.

data

A data.table containing the observed data.

confounders

A character vector of column names corresponding to baseline covariates.

modifiers

A character vector of columns names corresponding to the suspected effect modifiers. This vector must be a subset of confounders.

exposure

A character corresponding to the exposure variable.

outcome

A character corresponding to the outcome variable.

censoring

A character indicating the right censoring indicator variable. Only used with time-to-event outcomes. Defaults to NULL.

outcome_type

A character indicating the outcome type. "continuous", "binary" and "time-to-event" are currently supported.

effect

A character indicating the type of treatment effect modifier variable importance parameter. Currently supports "absolute" and "relative".

estimator

A character set to either "tmle" or "onestep". The former results in unihtee() to use a targeted maximum likelihood estimators to estimate the deisred TEM-VIP, while the latter uses a one step estimator.

cond_outcome_estimator

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the conditional outcome. Defaults to a generalized linear model with one- and two- way interactions among all confounders and exposure variables. Only used with continuous and binary outcomes.

prop_score_estimator

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the propensity score. Defaults to a generalized linear model with one- and two- way interactions among all confounders variables.

prop_score_values

An optional character corresponding to the (known) propensity score values for each observation in data. Defaults to NULL.

failure_hazard_estimator

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the conditional failure hazard function. Defaults to an xgboost learner with confounders and exposure variables as covariates. Only used with time-to-event outcomes.

censoring_hazard_estimator

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the conditional censoring hazard function. Defaults to an xgboost learner with confounders and exposure variables as covariates. Only used with time-to-event outcomes.

Value

A list object containing the validation dataset's uncentered efficient influence function estimates, the treatment effect modification variable importance parameter estimates and the proportion of observations in the validation data.