Conditional Outcome Estimator

fit_cond_outcome() estimates the conditional outcome, based on the exposure and confounders, over the train_data dataset. The estimator used for this estimation is based on the learners argument, and the covariates considered are specified by the confounders and exposure arguments.

Usage

fit_cond_outcome(
  train_data,
  valid_data,
  learners,
  outcome,
  exposure,
  confounders
)

Arguments

train_data

A data.table containing the observed data. train_data is formatted by unihtee().

valid_data

An optional data.table representing a holdout dataset of the observed data. It is only used for cross-fitting purposes. Defaults to NULL.

learners

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the propensity score model.

outcome

A character corresponding to the outcome variable.

exposure

A character corresponding to the exposure variable.

confounders

A character vector of column names corresponding to baseline covariates.

Value

A named list of four elements. (1) "estimates", the expected conditional outcome for each observation in valid_data, if specified, or train_data otherwise. (2) "fit", the trained Stack or learner. (3) "exp_estimates", the expected conditional outcome for each observation in valid_data, if specified, or train_data otherwise, had these observations been exposed. (4) "noexp_estimates", the expected conditional outcome for each observation in valid_data, if specified, or train_data

otherwise, had these observations not been exposed.