Propensity Score Estimator

fit_prop_score() estimates the propensity score over the train_data dataset. The estimator used for this estimation is based on the learners argument, and the covariates considered are specified by the confounders argument.

Usage

fit_prop_score(train_data, valid_data, learners, exposure, confounders)

Arguments

train_data

A data.table containing the observed data. train_data is formatted by unihtee().

valid_data

An optional data.table representing a holdout dataset of the observed data. It is only used for cross-fitting purposes. Defaults to NULL.

learners

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the propensity score model.

exposure

A character corresponding to the exposure variable.

confounders

A character vector of column names corresponding to baseline covariates.

Value

A named list of two elements. (1) "estimates", the propensity score estimates for each observation in valid_data, if specified, or train_data otherwise. (2) "fit", the trained Stack or learner.