One-Step Estimator of the Marginal Relative Effect for Time-to-Event Outcomes

One-Step Estimator of the Marginal Relative Effect for Time-to-Event Outcomes

Usage

one_step_estimator_rmst_log_outcome(
  data,
  exposure,
  outcome,
  prop_score_fit,
  prop_score_values,
  failure_hazard_fit,
  censoring_hazard_fit
)

Arguments

data

A data.table containing the observed data. train_data is formatted by unihtee().

exposure

A character corresponding to the exposure variable.

outcome

A character corresponding to the outcome variable.

prop_score_fit

A list output by the fit_prop_score() function.

prop_score_values

A numeric vector corresponding to the (known) propensity score values for each observation in data.

failure_hazard_fit

A list output by the fit_cond_outcome() function.'

censoring_hazard_fit

A list output by the fit_censoring_hazard() function.'

Value

The one-step estimate.