Plot a unihtee Object

plot.unihtee() produces a plot depicting the estimated simple linear regression line for the specified modifier parameter estimated by unihtee().

Usage

# S3 method for class 'unihtee'
plot(x, y, ..., modifier_name, print_interpretation = TRUE)

Arguments

x

A unihtee class object output by unihtee().

y

Ignored.

...

Ignored.

modifier_name

A character specifying the treatment effect modifier variable importance parameter estimate to plot.

print_interpretation

A flag indicating whether to print the interpretation of the TEM-VIP inference plot. Defaults to TRUE.

Value

A ggplot2 object.