Plug-In Estimator

plugin_estimator() estimates the treatment effect modifier variable importance parameters using a plug-in estimation strategy.

Usage

plugin_estimator(
  data,
  outcome,
  modifiers,
  effect,
  cond_outcome_fit,
  failure_hazard_fit
)

Arguments

data

A data.table containing the observed data. train_data is formatted by unihtee().

outcome

A character corresponding to the outcome variable.

modifiers

A character vector of columns names corresponding to the suspected effect modifiers. This vector must be a subset of confounders.

effect

A character indicating the type of treatment effect modifier variable importance parameter. Currently supports "absolute" and "relative".

cond_outcome_fit

A list output by the fit_cond_outcome() function.

failure_hazard_fit

A list output by the fit_cond_outcome() function.'

Value

A vector containing the plug-in estimates for each potential modifier.