Time-to-Event Data Melt

tte_data_melt() turns wide-format data.table objects into long-format data.table objects.

Usage

tte_data_melt(
  data,
  confounders,
  exposure,
  outcome,
  censoring,
  time_cutoff,
  prop_score_values
)

Arguments

data

A data.table containing the observed data.

confounders

A character vector of column names corresponding to baseline covariates.

exposure

A character corresponding to the exposure variable.

outcome

A character corresponding to the outcome variable.

censoring

A character indicating the right censoring indicator variable.

time_cutoff

A numeric representing the time point at which to evaluate the time-to-event parameter.

prop_score_values

A character corresponding to the (known) propensity score values for each observation in data.

Value

A long, time-to-event version of the original data.table object. That is, observations are repeated at time points between the earliest time in the data and the time_cutoff argument.