Univariate Heterogeneous Treatment Effect Modifier Estimator

unihtee() estimates treatment effect modifiers variable importance parameters (TEM-VIPs). These TEM-VIPs are defined on the additive or relative effects and can be estimated using one-step or targeted maximum likelihood estimators.

unihtee(
  data,
  confounders,
  modifiers,
  exposure,
  outcome,
  censoring = NULL,
  time_cutoff = NULL,
  outcome_type = c("continuous", "binary", "time-to-event"),
  effect = c("additive", "relative"),
  estimator = c("tmle", "onestep"),
  cross_fit = FALSE,
  cross_fit_folds = 5,
  cond_outcome_estimator = sl3::Lrnr_glm_fast$new(),
  prop_score_estimator = sl3::Lrnr_glm_fast$new(),
  prop_score_values = NULL,
  failure_hazard_estimator = sl3::Lrnr_xgboost$new(),
  censoring_hazard_estimator = sl3::Lrnr_xgboost$new(),
  parallel = FALSE
)

Arguments

data

A data.table containing the observed data.

confounders

A character vector of column names corresponding to baseline covariates.

modifiers

A character vector of columns names corresponding to the suspected effect modifiers. This vector must be a subset of confounders.

exposure

A character corresponding to the exposure variable.

outcome

A character corresponding to the outcome variable.

censoring

A character indicating the right censoring indicator variable. Only used with time-to-event outcomes. Defaults to NULL.

time_cutoff

A numeric representing the timepoint at which to evaluate the time-to-event parameter. Only used with time-to-event outcomes. Defaults to NULL.

outcome_type

A character indicating the outcome type. "continuous", "binary" and "time-to-event" are currently supported.

effect

A character indicating the type of treatment effect modifier variable importance parameter. Currently supports "additive" and "relative".

estimator

A character set to either "tmle" or "onestep". The former results in unihtee() to use a targeted maximum likelihood estimators to estimate the deisred TEM-VIP, while the latter uses a one step estimator.

cross_fit

A logical determining whether cross-fitting should be used. Defaults to FALSE.

cross_fit_folds

A numeric stating the number of folds to use in the cross-fitting procedure. Defaults to 5.

cond_outcome_estimator

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the conditional outcome. Defaults to a generalized linear model with one- and two- way interactions among all confounders and exposure variables. Only used with continuous and binary outcomes.

prop_score_estimator

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the propensity score. Defaults to a generalized linear model with one- and two- way interactions among all confounders variables.

prop_score_values

An optional character corresponding to the (known) propensity score values for each observation in data. Defaults to NULL.

failure_hazard_estimator

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the conditional failure hazard function. Defaults to an xgboost learner with confounders and exposure variables as covariates. Only used with time-to-event outcomes.

censoring_hazard_estimator

A Stack, or other learner class (inheriting from Lrnr_base), containing a set of learners from sl3 to estimate the conditional censoring hazard function. Defaults to an xgboost learner with confounders and exposure variables as covariates. Only used with time-to-event outcomes.

parallel

A logical stating if origami's built-in parallelized cross-validation routines should be used when cross_fit = TRUE. The future suite is used. Defaults to FALSE.

Value

A data.table containing the effect estimates and (adjusted) p-values of the modifiers. The suspected treatment effect modifiers ordered according to ascending p-values.